+++ commander.h
@@ -24,7 +24,7 @@ void send_finish_command();
 void send_bearing_message(double &bearing, int32_t &lat, int32_t &lon, float &alt);
 
 /* send the rssi message */
-void send_rssi_message(int &rssi, float &heading, int32_t &lat, int32_t &lon, float &alt);
+void send_rssi_message(int &rssi, int16_t &heading, int32_t &lat, int32_t &lon, float &alt);
 
 int write_to_serial(mavlink_message_t &message);
